logo-formatter/pyproject.toml
[project]
name = "logo-formatter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pillow>=10.0.0",
]
